シェアボタンのコンポーネント化
